msmnile_au: Fix adb push into /vendor/firmware_mnt
adb is not able to push into /vendor/firmware_mnt even when
it is mounted as rw. The reason is that it is not able to
change user using chown. To fix this, either the fstab mount
options can be changed or config.fs can be modified. Changing
mount options to keep in sync with mobile.

Change-Id: Ia30c9f486517e273835eab2b951bf22b6ace9b9a