msmnile_au: Enable UClamp configuration
Schedtune has been removed from kernel5.4 and UClamp framework
must be used in it's place instead. This change configures
zygote to use the uclamp framework and does settings related to
it. Currently, those settings have been made for walt scheduler and
a different set of settings maybe required for different kernel version
or if walt is not used.

Change-Id: I58ae620000818356bbe2e13d36fdf4b200232bb1