msmnile_au: Remove the unconditional enabling of cnss2.ko
cnss2.ko was enabled unconditionally by the following change:

 >>> Change-Id: Icb3d4eb92f386541aed4a13514b1cb7b9c443262

This was done as part of the efforts to modularize pcie and wlan
platform drivers on Android R, and wasn't exposed as a problem
due to cnss2.ko was compiled as an in-tree module and has never
been disabled.

With Android T, cnss2.ko is compiled as an out-of-tree module
and shouldn't be included unconditionally, so remove it from
msmnile_au.mk.

Instead, the wlan platform driver will be included in wlan.mk
which is located in another component. The wlan.mk script will
be included based on the value defined for BOARD_HAS_QCOM_WLAN.

Change-Id: Iaf0b333e558ee4d01f0c918da7ef07069b4f0286