msmnile_au: launch ais_server in init
Fix: vendor.qti.automotive.qcarcam@1.0-service crash on boot is fixed.
ais_server fails as drivers are not loaded. Hence causes the qcarcam to restart.
The qcarcam service fails as Binder driver could not be opened in early-init phase.
This change would be reversed after the early service feature is implemented.