msmnile_au: Add support to modem during S2D
During Suspend-to-Disk (S2D), all subsystems—including ADSP, CDSP, and MODEM—must be powered off and subsequently powered on.

Support for ADSP and CDSP was previously added in the hibernation scripts. With the introduction of the MDF use case, the MODEM subsystem is now integrated into the hibernation flow. This patch updates the hibernation freeze and restore paths to properly manage the MODEM state during S2D.

Change-Id: I58a75f5604a622a51e54075504de6bc66be96467
Signed-off-by: Deepak Suresh <quic_deesur@quicinc.com>